fix: update all docs to have todd copywrite & expanded on file structure
