+++ eslint.config.mjs
@@ -1,4 +1,5 @@
 // Copyright Todd LLC, All rights reserved.
+
 import storybook from 'eslint-plugin-storybook';
 
 import { dirname } from 'path';
