+++ jest.setup.intl.js
@@ -1,4 +1,5 @@
 // Copyright Todd LLC, All rights reserved.
+
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 const enMessages = require('./src/messages/en.json');
