+++ next.config.ts
@@ -1,4 +1,5 @@
 // Copyright Todd LLC, All rights reserved.
+
 import type { NextConfig } from 'next';
 import createNextIntlPlugin from 'next-intl/plugin';
