+++ src/components/common/carousel/types/carousel.ts
@@ -1,3 +1,5 @@
+// Copyright Todd LLC, All rights reserved.
+
 import { ReactNode } from 'react';
 
 /**
