+++ src/components/common/index.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 export { default as Button } from './button/button';
 export { default as Carousel } from './carousel/carousel';
