+++ src/components/common/locale-switcher/locale-switcher.tsx
@@ -1,3 +1,5 @@
+// Copyright Todd LLC, All rights reserved.
+
 'use client';
 
 import { useLocale } from 'next-intl';
