+++ src/components/common/smooth-scroll/smooth-scroll.tsx
@@ -1,3 +1,5 @@
+// Copyright Todd LLC, All rights reserved.
+
 'use client';
 
 import React, { useEffect } from 'react';
