+++ src/components/landing/hero/hero.stories.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 import Hero from './hero';
