+++ src/components/landing/news-highlight-card/news-highlight-card.test.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import { screen, renderWithAct } from '@/test/test-utils';
 import NewsHighlightCard from './news-highlight-card';
