+++ src/components/landing/quote/quote.stories.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
