+++ src/components/landing/scroll-shrink-wrapper/scroll-shrink-wrapper.test.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import React from 'react';
 import { render, screen } from '@testing-library/react';
