+++ src/components/landing/scroll-shrink-wrapper/scroll-shrink-wrapper.tsx
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 'use client';
