+++ src/i18n/request.ts
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import { getRequestConfig } from 'next-intl/server';
 import { routing } from './config';
