+++ src/lib/metadata.ts
@@ -1,4 +1,4 @@
-//Copyright Todd LLC, All rights reserved.
+// Copyright Todd LLC, All rights reserved.
 
 import type { Metadata, Viewport } from 'next';
