+++ src/middleware.ts
@@ -1,4 +1,5 @@
 // Copyright Todd LLC, All rights reserved.
+
 import createMiddleware from 'next-intl/middleware';
 import { routing } from './i18n/config';
