+++ vitest.config.ts
@@ -1,4 +1,5 @@
 // Copyright Todd LLC, All rights reserved.
+
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
